pyproject.toml
[project]
name = "cassandra-asyncio-driver"
description = "Wrapper for cassandra-driver that makes queries awaitable using asyncio"
authors = [
    {name = "Mark Parncutt", email = "me@markparncutt.com"},
]
dynamic = ["version"]
dependencies = [
    "cassandra-driver==3.29.2"
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.version]
source = "scm"
write_to = "cassandra_asyncio/version.py"
write_template = "# Dynamically generated by build system, do not modify.\n__version__ = version = \"{}\"\n"

[tool.pdm.build]
excludes = ["./**/.git"]
includes = ["cassandra_asyncio/"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[dependency-groups]
test = ["pytest>=7.4.3"]
